Random savegame crash:
With all these saves in my nightly build save folder, the game crashes
when opening the "load game" dialog. It is not persistent, it happens
sometimes and other times it doesnt happen.

It seemed for a period of time to circle around the savegame called
"Gsö-Gsoe - Test.sve", which for a period could crash the game when it
was existing, but after a while it loaded fine.

Copy your existing savegames to a secure location and add all these
saves to the folder instead.